requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Pillow==10.0.0
numpy==1.24.3
python-dotenv==1.0.0
